refactor: remove the extra Root.env containers
Now we use dynamic ES6 `import()` to import one or another
component depending of the current environment.
This way, we also handle the Root in just one file, less code and
also a short file.

I also add here a `<App>` container. This component will host the
app routes and also will do the main fetching of data of the app
(data that need to be shared between several components). And,
obviously, that data fetched will be stored into Redux store, for
being required from everywhere inside the app.